Decompose FormContext into 3 single-responsibility classes (#981)
FormContext (542 lines) had dual responsibility: mutable state holder
and UI builder utility, with dimensional analysis as a third concern.

Split into:
- FormContext: mutable state only (~85 lines)
- FormFieldBuilder: UI field creation and layout (~270 lines)
- DimensionalAnalysisUI: equation validation and DA feedback (~170 lines)

Update all ElementForm implementations and PropertiesPanel to receive
the new collaborators via constructor injection.
